Moving tests into one file
